+++ changelog.md
@@ -2,13 +2,14 @@
 
   **Note: This update changed some config values, it is highly recommended to re-generate your config file!**
   * ADDED: Colored Lights support for the Colorful Lamp!
-  * ADDED: More Railcraft Integration
+  * ADDED: More Railcraft Integration (Requires Railcraft 9.4.x now)
     - Digital Signal Receiver Box: Same as the Receiver Box driver has been, but also fires an "aspect_changed" event whenever the box aspect changes
     - Digital Detector: Fires a "minecart" event whenever a minecart passes, containing the type of cart as well as the name if the cart has got one
     - Now you can get and set the force of Launcher Tracks using ComputerCraft or OpenComputers!
     - Now you can get and set the fuse time of Priming Tracks using ComputerCraft or OpenComputers!
     - Now you can check isPowered() on every kind of track that can be powered by Redstone
     - Replicated the Steam Turbine Driver and the Boiler Firebox Driver added by OpenComputers for ComputerCraft
+    - Added getDraw() for anything using Railcraft Charge
   * ADDED: Buildcraft Integration (Gate Statements)
     - 2 Triggers (Computer Running/Stopped) for the OpenComputers Computer Case
     - 2 Actions (Start/Stop Computer) for the OpenComputers Computer Case